refactor: update MessagesRepository usage and improve command descriptions
