+++ app/Console/Commands/Connectors/PromptMine/SendPushMonthlyPromptCountCommand.php
@@ -8,7 +8,7 @@ use Baka\Traits\KanvasJobsTrait;
 use Illuminate\Console\Command;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Connectors\PromptMine\Jobs\SendMonthlyMessageCountJob;
-use Kanvas\Social\Messages\Repositories\MessagesRepository;
+use Kanvas\Connectors\PromptMine\Repositories\MessagesRepository;
 use Kanvas\Social\MessagesTypes\Models\MessageType;
 use Kanvas\Users\Models\UsersAssociatedApps;
